Create separate bank & bowling folder structures
