+++ src/main/java/com/example/bank/BankService.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.bank;
 
 public interface BankService {
 	void pay(String id, double amount);
