+++ src/main/java/com/example/bank/Employee.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.bank;
 
 public class Employee {
