+++ src/main/java/com/example/bank/EmployeeRepository.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.bank;
 
 import java.util.List;
